refactor: move `ContextChatEngine` and `SimpleChatEngine`
