+++ packages/core/src/chat-engine/index.ts
@@ -1,36 +1,9 @@
-import type { ChatMessage, MessageContent } from "../llms";
-import type { BaseMemory } from "../memory";
-import { EngineResponse } from "../schema";
-
-export interface BaseChatEngineParams<
-  AdditionalMessageOptions extends object = object,
-> {
-  message: MessageContent;
-  /**
-   * Optional chat history if you want to customize the chat history.
-   */
-  chatHistory?:
-    | ChatMessage<AdditionalMessageOptions>[]
-    | BaseMemory<AdditionalMessageOptions>;
-}
-
-export interface StreamingChatEngineParams<
-  AdditionalMessageOptions extends object = object,
-> extends BaseChatEngineParams<AdditionalMessageOptions> {
-  stream: true;
-}
-
-export interface NonStreamingChatEngineParams<
-  AdditionalMessageOptions extends object = object,
-> extends BaseChatEngineParams<AdditionalMessageOptions> {
-  stream?: false;
-}
-
-export abstract class BaseChatEngine {
-  abstract chat(params: NonStreamingChatEngineParams): Promise<EngineResponse>;
-  abstract chat(
-    params: StreamingChatEngineParams,
-  ): Promise<AsyncIterable<EngineResponse>>;
-
-  abstract chatHistory: ChatMessage[] | Promise<ChatMessage[]>;
-}
+export {
+  BaseChatEngine,
+  type BaseChatEngineParams,
+  type NonStreamingChatEngineParams,
+  type StreamingChatEngineParams,
+} from "./base";
+export { ContextChatEngine } from "./context-chat-engine";
+export { DefaultContextGenerator } from "./default-context-generator";
+export { SimpleChatEngine } from "./simple-chat-engine";
