+++ packages/core/src/chat-engine/type.ts
@@ -1,10 +1,11 @@
-import type { ChatMessage } from "@llamaindex/core/llms";
-import type { NodeWithScore } from "@llamaindex/core/schema";
+import type { ChatMessage } from "../llms";
+import type { NodeWithScore } from "../schema";
 
 export interface Context {
   message: ChatMessage;
   nodes: NodeWithScore[];
 }
+
 /**
  * A ContextGenerator is used to generate a context based on a message's text content
  */
