+++ packages/llamaindex/src/engines/chat/index.ts
@@ -1,4 +1,2 @@
+export * from "@llamaindex/core/chat-engine";
 export { CondenseQuestionChatEngine } from "./CondenseQuestionChatEngine.js";
-export { ContextChatEngine } from "./ContextChatEngine.js";
-export { SimpleChatEngine } from "./SimpleChatEngine.js";
-export * from "./types.js";
